requirements.txt
pytest
wheel
torch==2.1.2
torchvision==0.16.2
transformers
tokenizers
numpy
opencv-python
supervision==0.19.0
openmim
mmcv-lite==2.0.1
mmdet==3.3.0
mmengine==0.10.4
mmcv==2.1.0
mmyolo
pillow
lvis-api